RecoveryCodestPage: refactored / recovery to backup naming refactor
