+++ packages/web-app/src/MobileRoutes.tsx
@@ -6,11 +6,11 @@ import { Redirect, Route, Switch, withRouter } from 'react-router'
 import { getStore } from './Store'
 import { MobilePageNotFound } from './components'
 import { MobileAccountSummaryContainer } from './modules/account-views-mobile'
+import { BackupCodesPageContainer } from './modules/backup-codes/BackupCodesPageContainer'
 import { MobileEarningSummaryContainer } from './modules/earn-views-mobile'
 import { PasskeyDeletePageContainer } from './modules/passkey-delete'
 import { PasskeySetupPageContainer, isPasskeyFeatureEnabled } from './modules/passkey-setup'
 import { PasskeySuccessPageContainer } from './modules/passkey-success'
-import { RecoveryCodesPageContainer } from './modules/recovery-codes'
 import { RewardDetailsContainer } from './modules/reward-views'
 
 class _Routes extends Component<RouteComponentProps> {
@@ -37,9 +37,7 @@ class _Routes extends Component<RouteComponentProps> {
           {isPasskeyFeatureEnabled && (
             <Route exact path="/account/passkey/delete/:id" component={PasskeyDeletePageContainer} />
           )}
-          {isPasskeyFeatureEnabled && (
-            <Route path="/account/backup-codes" exact component={RecoveryCodesPageContainer} />
-          )}
+          {isPasskeyFeatureEnabled && <Route exact path="/account/backup-codes" component={BackupCodesPageContainer} />}
           <Route component={MobilePageNotFound} />
         </Switch>
       </>
