+++ packages/web-app/src/modules/backup-codes/components/index.tsx
@@ -1,0 +1,1 @@
+export * from './BackupCodesPage'
