--- packages/web-app/src/modules/recovery-codes/components/index.tsx
@@ -1,1 +1,0 @@
-export * from './RecoveryCodesPage'
