--- packages/web-app/src/modules/recovery-codes/index.tsx
@@ -1,1 +1,0 @@
-export * from './RecoveryCodesPageContainer'
